% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundant_tax_physeq.R
\name{abundant_tax_physeq}
\alias{abundant_tax_physeq}
\title{Return top n most abundand taxonomic terms at a given level}
\usage{
abundant_tax_physeq(
  physeq = ps,
  lvl = "Class",
  top = 10,
  output_format = "ps",
  ignore_na = TRUE,
  silent = TRUE
)
}
\arguments{
\item{physeq}{a phyloseq object to rank}

\item{lvl}{a character vector defining the taxonomic level to rank}

\item{top}{an integer n defining the number of returned elements (e.g. 'top=5')}

\item{output_format}{"tops" or "ps" (default), depending on this value, function returns either a character vector or a pruned phyloseq object.}

\item{ignore_na}{TRUE/FALSE, default TRUE. If FALSE, includes 'NA' as a taxonomic term in the output.}

\item{silent}{TRUE/FALSE, default TRUE. If FALSE, prints the top n taxonomix terms as a message.}
}
\value{
a phyloseq object or character vector, depending on value of "output_format"
}
\description{
This function extracts the top n most abundant taxonomic terms at a given level.
It uses dplyr to calculate the most abundant elements.
Output_format will take "ps" or "tops" and either return the phyloseq object or a character vector of the n most common terms.
Ignores NA's by default but can show them (see parameters).
}
\examples{
ps_top10genera <- abundant_tax_physeq(physeq = ps, lvl="Genus", top=10)
top5species <- abundant_tax_physeq(physeq = ps, lvl="Species", top=5, output_format="tops", ignore_na = FALSE)

}
\author{
Simeon Rossmann
}
