% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_subset_ASVs.R
\name{list_subset_ASVs}
\alias{list_subset_ASVs}
\title{Returns a vector of ASVs belonging to groups at a taxonomic level}
\usage{
list_subset_ASVs(physeq = ps, subv = c("e"), taxlvlsub = "Kingdom")
}
\arguments{
\item{physeq}{a phyloseq object to extract ASVs/OTUs ('tax_names') from}

\item{subv}{a character vector of taxonomic terms at the level of interest}

\item{taxlvlsub}{a character string containing a taxonomic level present in the database}
}
\value{
a character vector of ASVs
}
\description{
A function to list ASVs belonging to a set of taxonomic terms
}
\examples{
phyt_ASVs <- list_subset_ASVs(physeq = ps, subv = c("Phytophthora","Phytopythium"), taxlvlsub="Genus")
ps_phyt <- phyloseq::prune_taxa(phyt_ASVs, ps)
}
\author{
Simeon Rossmann
}
