% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ranked_sums.R
\name{make_ranked_sums}
\alias{make_ranked_sums}
\title{Rank ASV abundance sums for a given phyloseq object}
\usage{
make_ranked_sums(physeq = ps, myset = "Not specified")
}
\arguments{
\item{physeq}{a phyloseq object to rank by abundance}

\item{myset}{optional: a character string describing the ranked physeq object.}
}
\value{
a tibble that contains all sample metadata (fcts), their abundance (dbl),
their ranks (int) by sum of abundance, their original order (int) and a custom descriptor (chr).
}
\description{
A function that ranks all samples in a phyloseq object by the sum of reads assigned to all ASVs in that sample (Abundance sum/Total ASV counts).
}
\examples{
ranked_controls <- make_ranked_sums(physeq = ps_contrls)

}
\author{
Simeon Rossmann
}
