% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_physeq.R
\name{summarise_physeq}
\alias{summarise_physeq}
\title{Summarize a phyloseq object}
\usage{
summarise_physeq(
  physeq = ps,
  ASV_sublist = c(),
  sublist_id = "sublist_id_not_defined",
  samp_names = TRUE
)
}
\arguments{
\item{physeq}{a phyloseq object to be summarized}

\item{ASV_sublist}{a character vector of ASVs/OTUs (phyloseq taxa). Empty by default.
If provided the summary will contain the percentage of abundance the sublist comprises.
Compatible with cuphyr::list_subset_ASVs().}

\item{sublist_id}{a character string describing the subset of taxa provided in ASV_sublist}

\item{samp_names}{TRUE/FALSE. If TRUE (default), sample names in the subset are included in the summary.}
}
\value{
a character string summarising the provided phyloseq object
}
\description{
Summarize the total number of ASVs, number of ASVs in a subset and percentage
of reads this ASV subset contains (of total) of a given phyloseq object.
}
\examples{
summarise_physeq(ps)
summarise_physeq(ps, ASV_sublist= c("ASV1"), sublist_id ="ASV1", samp_names=FALSE)

phyt_ASVs <- cuphyr::list_subset_ASVs(physeq = ps, subv = c("Phytophthora","Phytopythium"), taxlvlsub="Genus")
summarise_physeq(ps, ASV_sublist= phyt_ASVs, sublist_id ="Phytophthora and Phytopythium", samp_names=FALSE)

}
\author{
Simeon Rossmann
}
